Fix Goodix reset signal not working on some configuration.
Reset pin, at the end of power on sequence, was configured as input.
This choice was justified with the request to save power.
The problem is that it works only if there is a pull-up on the
reset line.
To avoid handling different cases, it was decided to leave reset pin
always configured as Output.