Fix DRM loop at boot when HDMI/MIPI/LVDS enabled
When panel_lvds driver is used as module, drm cannot bind
immediately the imx8mp_ldb and binding is deferred.
This happens only if the three outputs are enabled.

For some unknown reason, deferring binding does not work,
and kernel boot does not complete and drm try endless to bind ldb.
If panel_lvds is driver is built in, it doesn't happen.